+++ src/callbacks/postPlayerInit.ts
@@ -5,7 +5,12 @@ import * as nLuck from "../features/items/nLuck";
 import * as disableMultiplayer from "../features/mandatory/disableMultiplayer";
 
 export function main(player: EntityPlayer): void {
-  // log("MC_POST_PLAYER_INIT");
+  /*
+  const character = player.GetPlayerType();
+  log(
+    `MC_POST_PLAYER_INIT - character: ${character}, InitSeed: ${player.InitSeed}`,
+  );
+  */
 
   // Mandatory
   disableMultiplayer.postPlayerInit(player);